Move file logging to separate class
