+++ java/src/Ice/src/main/java/com/zeroc/Ice/LoggerI.java
@@ -5,7 +5,13 @@
 package com.zeroc.Ice;
 
 class LoggerI implements Logger {
-    LoggerI(String prefix, String file) {
+    private final String _prefix;
+    private final String _formattedPrefix;
+    private final String _lineSeparator;
+    private final java.text.DateFormat _date;
+    private final java.text.SimpleDateFormat _time;
+
+    LoggerI(String prefix) {
         _prefix = prefix;
 
         if (!prefix.isEmpty()) {
@@ -17,20 +23,6 @@ class LoggerI implements Logger {
         _lineSeparator = System.getProperty("line.separator");
         _date = java.text.DateFormat.getDateInstance(java.text.DateFormat.SHORT);
         _time = new java.text.SimpleDateFormat(" HH:mm:ss:SSS");
-
-        if (!file.isEmpty()) {
-            try {
-                _out = new java.io.FileOutputStream(new java.io.File(file), true);
-            } catch (java.io.FileNotFoundException ex) {
-                throw new InitializationException("FileLogger: cannot open " + file);
-            }
-        } else {
-            _out = null;
-        }
-    }
-
-    LoggerI(String prefix) {
-        this(prefix, "");
     }
 
     @Override
@@ -98,11 +90,15 @@ class LoggerI implements Logger {
 
     @Override
     public Logger cloneWithPrefix(String prefix) {
-        if (_out != null) {
-            throw new UnsupportedOperationException(
-                    "cloneWithPrefix is not supported on a file logger");
-        }
         return new LoggerI(prefix);
+    }
+
+    @Override
+    public void close() throws Exception {}
+
+    // Writes the actually message to the output stream.
+    protected void writeMessage(String message) {
+        System.err.print(message);
     }
 
     private void write(StringBuilder message, boolean indent) {
@@ -114,28 +110,6 @@ class LoggerI implements Logger {
             }
         }
         message.append(_lineSeparator);
-
-        if (_out == null) {
-            System.err.print(message.toString());
-        } else {
-            try {
-                _out.write(message.toString().getBytes());
-            } catch (java.io.IOException ex) {
-            }
-        }
+        writeMessage(message.toString());
     }
-
-    @Override
-    public void close() throws java.io.IOException {
-        if (_out != null) {
-            _out.close();
-        }
-    }
-
-    private final String _prefix;
-    private final String _formattedPrefix;
-    private final String _lineSeparator;
-    private final java.text.DateFormat _date;
-    private final java.text.SimpleDateFormat _time;
-    private final java.io.FileOutputStream _out;
 }
